# perf(npm): deserialize npm packument in thread pool