mac m2, 8 cores:

```
% hyperfine --warmup=3 "../../deno/deno_old run -A repro.ts" "../../deno/target/release/deno run -A repro.ts"
Benchmark 1: ../../deno/deno_old run -A repro.ts
  Time (mean ± σ):     457.4 ms ±   4.5 ms    [User: 363.7 ms, System: 103.0 ms]
  Range (min … max):   452.7 ms … 468.8 ms    10 runs
 
Benchmark 2: ../../deno/target/release/deno run -A repro.ts
  Time (mean ± σ):     403.2 ms ±   6.2 ms    [User: 385.1 ms, System: 111.5 ms]
  Range (min … max):   398.2 ms … 413.7 ms    10 runs
 
Summary
  ../../deno/target/release/deno run -A repro.ts ran
    1.13 ± 0.02 times faster than ../../deno/deno_old run -A repro.ts
```

```
% hyperfine --warmup=3 --prepare="rm -rf node_modules" "DENO_FUTURE=1 ../deno/deno_old i" "DENO_FUTURE=1 ../deno/target/release/deno i"
Benchmark 1: DENO_FUTURE=1 ../deno/deno_old i
  Time (mean ± σ):      1.672 s ±  0.030 s    [User: 0.503 s, System: 0.663 s]
  Range (min … max):    1.624 s …  1.722 s    10 runs
 
Benchmark 2: DENO_FUTURE=1 ../deno/target/release/deno i
  Time (mean ± σ):      1.300 s ±  0.013 s    [User: 0.624 s, System: 0.685 s]
  Range (min … max):    1.283 s …  1.320 s    10 runs
 
Summary
  DENO_FUTURE=1 ../deno/target/release/deno i ran
    1.29 ± 0.03 times faster than DENO_FUTURE=1 ../deno/deno_old i
```