@dsherret 2024-05-28
Closing in favour of https://github.com/denoland/deno/pull/24018/files